requirements.txt
flask
Flask-SQLAlchemy
flask-login
python-binance
numpy
newtulipy
websocket-client